@@ -25,27 +25,27 @@
     },
     "devDependencies": {
         "@angular-material-extensions/select-country": "^8.1.0",
-        "@angular-eslint/eslint-plugin-template": "15.2.1",
-        "@angular/platform-browser-dynamic": "~15.2.8",
-        "@angular/material-moment-adapter": "^15.2.8",
-        "@angular-devkit/build-webpack": "^0.1502.7",
-        "@angular-eslint/template-parser": "15.2.1",
-        "@angular-devkit/build-angular": "^15.2.7",
-        "@angular-eslint/eslint-plugin": "15.2.1",
-        "@angular-eslint/schematics": "15.2.1",
-        "@angular/platform-browser": "~15.2.8",
-        "@angular-eslint/builder": "15.2.1",
-        "@angular/compiler-cli": "~15.2.8",
-        "@angular/animations": "~15.2.8",
-        "@schematics/angular": "^15.2.7",
-        "@angular/compiler": "~15.2.8",
-        "@angular/material": "^15.2.8",
-        "@angular/router": "~15.2.8",
-        "@angular/common": "~15.2.8",
-        "@angular/forms": "~15.2.8",
-        "@angular/core": "~15.2.8",
-        "@angular/cdk": "^15.2.8",
-        "@angular/cli": "~15.2.7",
+        "@angular-eslint/eslint-plugin-template": "16.0.1",
+        "@angular/platform-browser-dynamic": "~16.0.0",
+        "@angular/material-moment-adapter": "^16.0.0",
+        "@angular-devkit/build-webpack": "^0.1600.0",
+        "@angular-eslint/template-parser": "16.0.1",
+        "@angular-devkit/build-angular": "^16.0.0",
+        "@angular-eslint/eslint-plugin": "16.0.1",
+        "@angular-eslint/schematics": "16.0.1",
+        "@angular/platform-browser": "~16.0.0",
+        "@angular-eslint/builder": "16.0.1",
+        "@angular/compiler-cli": "~16.0.0",
+        "@angular/animations": "~16.0.0",
+        "@schematics/angular": "^16.0.0",
+        "@angular/compiler": "~16.0.0",
+        "@angular/material": "^16.0.0",
+        "@angular/router": "~16.0.0",
+        "@angular/common": "~16.0.0",
+        "@angular/forms": "~16.0.0",
+        "@angular/core": "~16.0.0",
+        "@angular/cdk": "^16.0.0",
+        "@angular/cli": "~16.0.0",
         "@biesbjerg/ngx-translate-extract-marker": "^1.0.0",
         "@bartholomej/ngx-translate-extract": "^8.0.2",
         "@typescript-eslint/eslint-plugin": "5.59.2",
@@ -63,7 +63,7 @@
         "autoprefixer": "^10.4.14",
         "@types/node": "^18.16.3",
         "tailwindcss": "^3.3.2",
-        "typescript": "~4.9.5",
+        "typescript": "~5.0.4",
         "zone.js": "~0.13.0",
         "ts-node": "^10.9.1",
         "postcss": "^8.4.23",
